Initial commit
php script that loads a template xml (so that I don’t have to create
the whole thing from scratch), and then finds attributes that it needs
and modifies them. Then it writes a new xml to the server.

To do:
Check if file already exists.
Take input from the site.
comment further.