Getting the json in the fancy new html and js page
I’m going to use the JSON elements of expts_pretty.json to make the
form and then set a session array of strings (names of the streams) for
the php script to grab and compare to the big bad json of timstreams.
I might need the expts_pretty.json to include a list of the available
resolutions at some point down the track (but don’t worry about it now).