Mockup second experiment, jquery badassery.
I set up some jQueery and JS awesomeness with the experiment and time
cam selection stuff.
Ill try sticking everything into a table so that it looks nice.
TODO: update the php script to use session variables arrays of selected
timestreams instead of all the time streams

I could use a link to a thumbnail for each timecam soon, so I can move
past functionality and start making the things pretty. I’ll include
that kind of info in my next update of the json schema, but for now it
should be fine. I might even be able to write a php script to do the
work for us (Thats something that we need to check php versions for,
but its totally doable).

I still haven’t started on on the CSS yet, but I’ll do that once the JS
+ jQueery stuff is finished.