TAKE THAT!
I got the election working.
The main page needs a json of the experiments and their associated
timestreams, and the php script needs a json of all the timestreams
(can be just in the same folder as the script).
Together, the main page will generate a form from the json, the form
will hide/show the appropriate timestreams for each experiment,
selecting the streams and hitting submit will generate a config (with
the selected layout), and a timegraph player for the selected streams.
TODO:
1. try with license key and actual data on borevitzlab.anu.edu.au
2. globals.php for all the global variables.
3. Pretty formatting and thumbnails instead of/for the checkboxes
4. Learn and implement the HTML5 player, coz its pretty.

This is ready for demo given the specs for json on github.